requirements.txt
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
asgiref==3.6.0
asn1crypto==1.5.1
astroid==2.12.13
astropy==5.1.1
astroquery==0.4.6
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==2.0.1
Babel==2.10.3
backcall==0.2.0
Brotli==1.0.9
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.7
certifi==2020.12.5
cffi==1.15.1
comtypes==1.1.14
config==0.5.1
constantly==15.1.0
contextlib2==21.6.0
contourpy==1.0.6
cryptography==38.0.3
cssselect2==0.7.0
cssselect==1.2.0
cycler==0.11.0
Cython==0.29.32
daphne==3.0.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
distlib==0.3.6
distro==1.8.0
dj-database-url==1.2.0
dj-static==0.0.6
django-allauth==0.52.0
django-autoslug==1.9.8
django-bootstrap-modal-forms==2.2.0
django-bootstrap3==22.2
django-braces==1.15.0
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-cors-headers==4.2.0
django-countries==7.5
django-crispy-forms==1.14.0
django-debug-toolbar==3.8.1
django-embed-video==1.4.8
django-extensions==3.2.1
django-filter==22.1
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-import-export==3.0.2
django-js-asset==2.0.0
django-localflavor==3.1
django-memcache-status==2.2
django-on-heroku==1.1.2
django-parler==2.3
django-qr-code==3.0.0
django-redisboard==8.4.0
django-rosetta==0.9.8
django-storages==1.13.2
django-taggit==3.1.0
django-tinymce==3.6.1
django-versatileimagefield==2.2
django-widget-tweaks==1.4.12
django-xhtml2pdf==0.0.4
Django==4.1.4
djangorestframework==3.14.0
dnspython==2.2.1
docutils==0.19
easy-thumbnails==2.8.5
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.87.0
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.9.0
filetype==1.2.0
flatbuffers==22.12.6
flower==1.2.0
fonttools==4.38.0
fpdf==1.7.2
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
google-auth==2.15.0
google-pasta==0.2.0
googletrans==3.0.0
greenlet==2.0.1
grpcio==1.51.1
gTTS==2.2.4
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.7.0
hiredis==2.1.0
hpack==3.0.0
hstspreload==2022.11.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
humanize==4.4.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
ImageHash==4.3.1
imageio==2.23.0
importlib-metadata==5.0.0
jsonfield==3.1.0
jsonschema==4.17.0
kombu==5.2.4
lazy-object-proxy==1.9.0
libclang==14.0.6
libimagequant==2.17.0.0
libmagic==1.0
linecache2==1.0.0
llvmlite==0.39.1
lockfile==0.12.2
lxml==4.9.2
markdown-it-py==2.2.0
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.4
more-itertools==9.0.0
mpmath==1.2.1
msgpack==0.6.2
multidict==6.0.2
mysql-connector-python==8.0.32
mysql==0.0.3
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
oauthlib==3.2.2
odfpy==1.4.1
onnxruntime==1.13.1
openai==0.27.6
openpyxl==3.0.10
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==21.3
pafy==0.5.5
pefile==2022.5.30
pep517==0.13.0
pep8==1.7.1
pickleshare==0.7.5
Pillow==9.3.0
pip==21.3.1
pipdate==0.5.6
platformdirs==2.5.3
plyer==2.0.0
polib==1.1.1
pooch==1.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
Protego==0.2.1
protobuf==3.19.6
psutil==5.9.4
psycopg2-binary==2.9.5
psycopg2-pool==1.1
psycopg2==2.9.5
pure-eval==0.2.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
PyAudio==0.2.12
pybase64==1.2.3
pycodestyle==2.10.0
pycparser==2.20
pydantic==1.10.4
PyDispatcher==2.0.6
pydub==0.25.1
pydyf==0.5.0
pyerfa==2.0.0.1
Pygments==2.13.0
PyHamcrest==2.0.2
pyhanko-certvalidator==0.19.8
pyHanko==0.16.0
pyinstaller-hooks-contrib==2022.10
pyinstaller==5.5
PyJWT==2.6.0
pylibjpeg-libjpeg==1.3.2
pylint==2.15.9
PyMatting==1.1.8
pymongo==4.3.3
pyOpenSSL==23.0.0
pyparsing==3.0.9
pypdf2==3.0.1
PyPDF3==1.0.6
pyphen==0.13.2
pypiwin32==223
pypng==0.20220715.0
PyQRCode==1.2.1
PyQt5-Qt5==5.15.2
pyreadline3==3.4.1
pyrsistent==0.19.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-magic==0.4.27
python-memcached==1.59
python-multipart==0.0.5
python-stdnum==1.18
python3-openid==3.2.0
pytoml==0.1.21
pyttsx3==2.90
pytz-deprecation-shim==0.1.0.post0
pytz==2022.7.1
pyvo==1.4
PyWavelets==1.4.1
pywin32-ctypes==0.2.0
pywin32==304
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qrcode==7.3.1
qtconsole==5.4.0
QtPy==2.3.0
queuelib==1.6.2
razorpay==1.3.0
redis==4.4.0
redisboard==0.0
regex==2022.10.31
requests-file==1.5.1
requests-oauthlib==1.3.1
requests==2.28.1
retrying==1.3.4
rfc3986==1.5.0
rich==13.3.5
rsa==4.9
s3transfer==0.6.0
segno==1.5.2
Send2Trash==1.8.0
service-identity==21.1.0
setuptools==60.2.0
simpervisor==0.4
simplejson==3.18.4
six==1.15.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.44
sqlparse==0.4.1
stack-data==0.6.0
starlette==0.21.0
static3==0.7.0
stripe==5.0.0
svglib==1.4.1
sympy==1.11.1
tablib==3.3.0
termcolor==2.1.1
terminado==0.17.0
textblob==0.17.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.1
traceback2==1.4.0
traitlets==5.5.0
twilio==7.16.0
twisted-iocpsupport==1.0.2
Twisted==22.10.0
txaio==22.2.1
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritools==4.0.0
urllib3==1.25.11
uvicorn==0.20.0
vine==5.0.0
vis==0.0.5
w3lib==2.1.1
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wheel==0.37.1
whitenoise==6.2.0
widgetsnbextension==4.0.3
wrapt==1.14.1

